+++ processes/ProcessManager.h
@@ -2,6 +2,7 @@
 #include <memory>
 #include <list>
 #include <processes/Process.h>
+
 namespace gamelib
 {
 	class ProcessManager
@@ -14,7 +15,7 @@ namespace gamelib
 		std::weak_ptr<Process> AttachProcess(const std::shared_ptr<Process>& process);
 		void AbortAllProcesses(bool immediate) const { for (auto& process : processes) process->OnAbort(); }
 
-		[[nodiscard]] unsigned int GetProcessCount() const { return processes.size(); }
+		[[nodiscard]] unsigned int GetProcessCount() const { return static_cast<unsigned int>(processes.size()); }
 	private:
 
 		std::list <std::shared_ptr<Process>> processes;
